FIX: disallow dice to be held after tenzies
Add gameInProgress in replace of firstStart to track if the game is in progress. If it is the first render, gameInProgress and tenzies  should be false, which can be used to check if it is the first start. Add a parameter that represents the active status of dice to generateNewDice and allNewDice to improve flexibility. When game is first rendered, a set of inactive dice will be generated so that user won't trigger any effect by clicking on it. After tenzies, the dice will be set to be inactive so user can't unhold them before they start a new game.